requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
line-bot-sdk==3.5.0
Pillow==10.1.0
pytesseract==0.3.10
python-dotenv==1.0.0
requests==2.31.0
python-multipart==0.0.6
opencv-python
numpy==1.24.3
gunicorn==21.2.0
easyocr==1.7.0
torch==2.1.0
torchvision==0.16.0
opencv-python==4.8.1.78